requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
groq==0.4.2
sqlite3-python==1.0.0
pydantic==2.5.0
colorlog==6.7.0
python-dateutil==2.8.2
pytz==2023.3
